Assistant checkpoint: Refactor ExportFacadeFactory into interface with implementations
Assistant generated file changes:
- src/main/java/org/joshsim/lang/export/ExportFacadeFactory.java: Convert to interface
- src/main/java/org/joshsim/lang/export/JvmExportFacadeFactory.java: Create JVM implementation
- src/main/java/org/joshsim/lang/export/WasmExportFacadeFactory.java: Create WASM implementation

---

User prompt:

Hello! I would like to change @src/main/java/org/joshsim/lang/export/ExportFacadeFactory.java into an interface, keeping the JavaDoc in place but moving the logic to an implementor at @src/main/java/org/joshsim/lang/export/JvmExportFacadeFactory.java. Can you please perform that refactor, adding JavaDoc to the top of @src/main/java/org/joshsim/lang/export/JvmExportFacadeFactory.java And to the top of the class but not adding JavaDoc to "@Override" methods? Then, can you please add an additional implementor in @src/main/java/org/joshsim/lang/export/WasmExportFacadeFactory.java With JavaDoc at the top of the file and class but with exceptions thrown in the methods indicating that they are not yet implemented? Thanks!
